refactor: Merge types/docs from testing/auth/api_key_usage_tracker.pyi
- Merged type hints and docstrings from crudclient/testing/auth/api_key_usage_tracker.pyi into crudclient/testing/auth/api_key_usage_tracker.py.

- Deleted crudclient/testing/auth/api_key_usage_tracker.pyi.
